requirements.txt
streamlit==1.33.0
pandas==2.2.2
numpy==1.26.4
google-generativeai==0.3.2
python-dotenv==1.0.1
plotly==5.22.0
openpyxl==3.1.2
protobuf==3.20.3
textblob==0.17.1
scikit-learn==1.4.2
networkx==3.2.1
matplotlib==3.8.4
fpdf2==2.8.3  
streamlit-authenticator==0.2.3
sqlalchemy==2.0.30
pandasai>=2.0